requirements.txt
fastapi
pydantic
pillow
requests
uvicorn
google-generativeai
fastapi-cors
python-multipart
